Safari-611.1.10/LayoutTests/js/dom/date-big-setdate-expected.txt

This test checks for regression against: 3381 Date.prototype.setDate() incorrect for values >=128.
12975: DST changes in US affect JavaScript date calculations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS d.valueOf() - curValue is millisecondsPerDay
PASS Passed all tests for DST (or skipped the tests if your timezone isn't PST/PDT)
PASS successfullyParsed is true

TEST COMPLETE
